requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
decouple==0.0.7
Django==4.2.8
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
kafka-python==2.0.2
Pillow==10.1.0
psycopg2-binary==2.9.9
PyJWT==1.7.1
python-decouple==3.8
pytz==2023.3.post1
rest-framework-simplejwt==0.0.2
sqlparse==0.4.4
typing-extensions==4.9.0
confluent-kafka==1.7.0
